rename github utils to changelog
